+++ packages/cli-plugin-vercel/src/utils.ts
@@ -1,0 +1,13 @@
+export const convertMethods = (method: string | string[]) => {
+  if (!method || !method.length) {
+    return;
+  }
+  const methods = [].concat(method);
+  const upperMthoeds = methods.map(method => {
+    return method.toUpperCase();
+  });
+  if (upperMthoeds.includes('ANY') || upperMthoeds.includes('ALL')) {
+    return;
+  }
+  return upperMthoeds;
+};